feat(top-bar): re-arrange folders, use sass, globalize container component
