--- components/_temp-styles.module.css
@@ -1,13 +1,0 @@
-.header {
-  max-width: 80%;
-  margin: 0 auto;
-  height: 80px;
-  display: flex;
-  align-items: center;
-}
-.header a {
-  margin-right: 3rem;
-}
-.header a:nth-child(2) {
-  margin-left: auto;
-}
