+++ components/container/index.js
@@ -1,0 +1,7 @@
+import styles from './styles.module.scss'
+
+const Container = ({ children }) => {
+  return <div className={styles.container}>{children}</div>
+}
+
+export default Container
