+++ components/container/styles.module.scss
@@ -1,0 +1,5 @@
+.container {
+  max-width: 80%;
+  color: #333;
+  margin: 0 auto;
+}
