+++ package.json
@@ -18,10 +18,12 @@
   "dependencies": {
     "@tinacms/api-git": "^0.7.0",
     "@tinacms/git-client": "^0.6.2",
+    "@zeit/next-sass": "^1.0.1",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "moment": "^2.24.0",
     "next": "^9.3.1",
+    "node-sass": "^4.13.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "styled-components": "^5.0.1",
