+++ pages/features.js
@@ -1,7 +1,6 @@
-import styled from "styled-components"
-
 import Head from "../components/head"
 import Layout from "../components/layout"
+import Container from "../components/container"
 
 const Features = () => {
   return (
@@ -13,11 +12,5 @@ const Features = () => {
     </Layout>
   )
 }
-
-const Container = styled.div`
-  max-width: 80%;
-  color: #333;
-  margin: 0 auto;
-`
 
 export default Features
